% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{add}
\alias{add}
\title{Basic computations}
\usage{
add(x, y)
}
\arguments{
\item{x}{Numeric.}

\item{y}{Numeric.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Basic computations
}
\examples{
add(4,8)
}
