% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPackage-package.R
\docType{package}
\name{ExPackage-package}
\alias{ExPackage}
\alias{ExPackage-package}
\title{ExPackage: This is a Demo}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Pierre Barbillon \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item collectif
}

}
\keyword{internal}
